+++ UI/UIUtils.cs
@@ -12,7 +12,7 @@ public static class UIUtils {
     /// <param name="baseObject">The base object for the path</param>
     /// <param name="path">The object path going from the base object</param>
     /// <returns></returns>
-    public static GameObject FindRecursive(this GameObject baseObject, string path) {
+    public static GameObject? FindRecursive(this GameObject baseObject, string path) {
         Transform t = baseObject.transform;
         string[] pathItems = path.Split("/");
         for (int i = 0; i < pathItems.Length; i++) {
@@ -31,8 +31,9 @@ public static class UIUtils {
     /// Pretty much GameObject.Find but also considers inactive objects and works for nested items
     /// </summary>
     /// <param name="path">The object path</param>
+    /// <param name="warnings">Whether to warn when the target is not found. Set to false when you know it might be not found.</param>
     /// <returns></returns>
-    public static GameObject FindRecursive(string path) {
+    public static GameObject? FindRecursive(string path) {
         int slashIndex = path.IndexOf('/');
         string firstItem = "";
         string restPath = "";
@@ -56,5 +57,53 @@ public static class UIUtils {
     /// <param name="uiObject">The GameObject to update</param>
     public static void UnfuckLayoutHack(this GameObject uiObject) {
         LayoutRebuilder.ForceRebuildLayoutImmediate((RectTransform)uiObject.transform);
+    }
+
+    // NEW WARNING PARAM VARIANT AS OVERLOAD TO AVOID BREAKAGE
+
+    /// <summary>
+    /// Pretty much GameObject.Find but also considers inactive objects and works for nested items
+    /// </summary>
+    /// <param name="baseObject">The base object for the path</param>
+    /// <param name="path">The object path going from the base object</param>
+    /// <param name="warnings">Whether to warn when the target is not found. Set to false when you know it might be not found.</param>
+    /// <returns></returns>
+    public static GameObject? FindRecursive(this GameObject baseObject, string path, bool warnings = true) {
+        Transform t = baseObject.transform;
+        string[] pathItems = path.Split("/");
+        for (int i = 0; i < pathItems.Length; i++) {
+            string itemStr = pathItems[i];
+            t = t.transform.Find(itemStr);
+            if (t == null) {
+                if (warnings) Plugin.Log.LogWarning($"{itemStr} not found for object path {baseObject.name}/{path}");
+                return null;
+            }
+        }
+
+        return t.gameObject;
+    }
+
+    /// <summary>
+    /// Pretty much GameObject.Find but also considers inactive objects and works for nested items
+    /// </summary>
+    /// <param name="path">The object path</param>
+    /// <param name="warnings">Whether to warn when the target is not found. Set to false when you know it might be not found.</param>
+    /// <returns></returns>
+    public static GameObject? FindRecursive(string path, bool warnings = true) {
+        int slashIndex = path.IndexOf('/');
+        string firstItem = "";
+        string restPath = "";
+        if (slashIndex != -1) {
+            firstItem = path.Substring(0, slashIndex);
+            restPath = path.Substring(slashIndex + 1);
+        }
+
+        GameObject baseObject = SceneManager.GetActiveScene().GetRootGameObjects()
+            .Where(obj => obj.name == firstItem).FirstOrDefault();
+        if (baseObject == null) {
+            Plugin.Log.LogWarning($"Root item not found for object path {path}");
+            return null;
+        }
+        return FindRecursive(baseObject, restPath, warnings: warnings);
     }
 }